rename testscommon to avoid package name collision with 'tests'
